requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
alembic>=1.14.0,<2.0.0
psycopg2-binary>=2.9.10,<3.0.0
pydantic-settings>=2.7.1,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.20,<1.0.0
email-validator>=2.2.0,<3.0.0
redis>=5.2.1,<6.0.0
celery>=5.4.0,<6.0.0
boto3>=1.35.99,<2.0.0
jinja2>=3.1.5,<4.0.0
fpdf2>=2.8.2,<3.0.0
httpx>=0.28.1,<1.0.0
pytest>=8.3.4,<9.0.0
pytest-asyncio>=0.25.0,<1.0.0
pytest-cov>=6.0.0,<7.0.0
faker>=33.1.0,<34.0.0
